+++ NEWS.md
@@ -1,5 +1,12 @@
 ## rtables 0.6.11.9001
 
+### New Features
+ * Added `stat_string` to `as_result_df(make_ard = TRUE)` to preserve the original string representation of the statistics.
+ * Added `add_tbl_name_split` to `as_result_df()` to handle split levels constituted by different table names.
+ 
+### Bug Fixes
+ * Fixed issue with `split_cols_by_multivar()` when having more than one value. Now `as_result_df(make_ard = TRUE)` adds a predefined split name for each of the `multivar` splits.
+
 ## rtables 0.6.11
 
 ### New Features